+++ include/DJConnectApp.h
@@ -350,6 +350,7 @@ private:
   uint32_t lastHaStatusAt_ = 0;
   uint32_t haPairingStartedAt_ = 0;
   uint32_t lastHaPairingScreenAt_ = 0;
+  uint32_t lastWakeWordGateLogAt_ = 0;
   uint32_t djResponseOverlayUntil_ = 0;
   uint32_t loopMetricsWindowStartedAt_ = 0;
   uint32_t loopMetricsBusyMs_ = 0;